+++ BodyweightFitness/WorkoutLog/WorkoutLog.storyboard
@@ -20,31 +20,36 @@
                         <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
                         <viewLayoutGuide key="safeArea" id="lTI-6J-eIx"/>
                     </view>
-                    <navigationItem key="navigationItem" id="Ftb-rR-i4s"/>
+                    <navigationItem key="navigationItem" id="Ftb-rR-i4s">
+                        <barButtonItem key="backBarButtonItem" title="Back" id="3nd-Hx-uZ8"/>
+                    </navigationItem>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="LQj-qE-ypR" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="159.19999999999999" y="128.18590704647679"/>
+            <point key="canvasLocation" x="854" y="127"/>
         </scene>
-        <!--Workout Log Navigation Controller-->
-        <scene sceneID="aLX-Gz-bJ1">
+        <!--Navigation Controller-->
+        <scene sceneID="NMJ-ao-xu3">
             <objects>
-                <navigationController storyboardIdentifier="WorkoutLogNavigationController" automaticallyAdjustsScrollViewInsets="NO" useStoryboardIdentifierAsRestorationIdentifier="YES" id="VPB-Nf-29m" customClass="WorkoutLogNavigationController" customModule="Bodyweight_Fitness" customModuleProvider="target" sceneMemberID="viewController">
-                    <extendedEdge key="edgesForExtendedLayout" top="YES"/>
+                <navigationController automaticallyAdjustsScrollViewInsets="NO" id="24k-Ia-upW" sceneMemberID="viewController">
                     <toolbarItems/>
-                    <navigationBar key="navigationBar" contentMode="scaleToFill" insetsLayoutMarginsFromSafeArea="NO" id="Ej0-Ay-SqI">
+                    <navigationBar key="navigationBar" contentMode="scaleToFill" insetsLayoutMarginsFromSafeArea="NO" id="jZh-Nd-HRY">
                         <rect key="frame" x="0.0" y="20" width="375" height="44"/>
                         <autoresizingMask key="autoresizingMask"/>
+                        <color key="tintColor" red="0.0" green="0.27058823529999998" blue="0.24313725489999999" alpha="1" colorSpace="calibratedRGB"/>
                         <color key="barTintColor" red="0.0" green="0.58823529409999997" blue="0.53333333329999999" alpha="1" colorSpace="calibratedRGB"/>
+                        <textAttributes key="titleTextAttributes">
+                            <color key="textColor" red="0.0" green="0.27058823529999998" blue="0.24313725489999999" alpha="1" colorSpace="calibratedRGB"/>
+                        </textAttributes>
                     </navigationBar>
                     <nil name="viewControllers"/>
                     <connections>
-                        <segue destination="TzD-qQ-pWO" kind="relationship" relationship="rootViewController" id="Rdf-mR-idf"/>
+                        <segue destination="TzD-qQ-pWO" kind="relationship" relationship="rootViewController" id="JUZ-b1-guQ"/>
                     </connections>
                 </navigationController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="OiJ-Ae-hDH" userLabel="First Responder" sceneMemberID="firstResponder"/>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="gZh-qr-8xp" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="-623" y="128"/>
+            <point key="canvasLocation" x="158.40000000000001" y="127.736131934033"/>
         </scene>
     </scenes>
 </document>
